General.jl documentation

General.jl documentation

Dynamical objects

Multiplex networks

Network models

Degree distributions

Graph IO

Simplicial complexes

Dynamical objects

General.count_states — Method

count_states(es, i, s)

Compute the number of edges in es which have state s at index i.

Examples

julia> count_states([[1,3], [3,3],[1,1]], 2, 3)
2

General.Ie! — Method

Internal dynamics for an identity StaticEdge. Copies the contents of v_s to e.

See also IStaticEdge

General.IStaticEdge — Method

Constructor for an identity StaticEdgewith dimension dim.

See also Ie!, SStaticEdge

General.SStaticEdge — Method

SStaticEdge(selection; invalid=0)

Constructor for a StaticEdge with selection dynamics. The edge has dimension equal to length(selection). After evaluation, e[i]==v_s[i] if selection[i]==true and else e[i]==invalid.

See also combine_graphs, IStaticEdge

General.majority_vertex — Method

majority_vertex(i::Integer, dim::Integer; U=1, A=2)

Construct a ODEVertex with binary Watts/threshold dynamics.

Arguments

i::Integer: index of the relevant state

dim::Integer: dimension of the total state

S1=1: value of state 1

S2=2: value of state two

Multiplex networks

General.combine_graphs — Method

static_edges, combined_graph = combine_graphs(gs::AbstractGraph...[; dims=[1], invalid=0])

Generate the combined_graph which edges are given by the union of the edges of gs. The goal of this function is offer an interface that allows to use multiplex networks in network_dynamics. Each argument represents a single layer. The array static_edges consists of SStaticEdges. The named keyword dims gives the dimension the StaticEdge in each layer ought to have. If length(dims)==1, all dimensions are equal to dim[1]. The selection of static_edges[i] is given by

[[edge ∈ edges(gs[i]) for j in 1:dims[i]] for i in eachindex(gs)]

The output of the StaticEdge dynamics corresponding to non-existing edges is set to invalid.

See also SStaticEdge

General.combine_graphs — Method

ode_vertices, static_edges, combined_graph =
    combine_graphs(v_f::Function, gs::AbstractGraph...[; dims=[1]])

Generate the combined_graph which edges are given by the union of the edges of gs. The goal of this function is offer an interface that allows to use multiplex networks in network_dynamics. Each argument represents a single layer. This method ought to bes used when the vertex dynamics depend on the degree in each layer.

Arguments

v_f::Function: a constructor ds -> ODEVertex, where ds[i] is the degree of the corresponding

vertex in gs[i].

gs::AbstractGraph...: graphs to combine

See also SStaticEdge

Network models

General.configuration_model — Method

configuration_model(degree_seq; allow_collisions=true, max_tries=1000)

Contruct a graph with degree distribution degree_seq using the configuration model. The implementation is based on that of the NetworkX python package.

Arguments

degree_seq::AbstractVector{<:Integer}: degree sequence of the graph

allow_collisions::Bool: if true, ignore multiple edges between the same pair of nodes and self-loops

max_tries::Integer: maximum number of attempts to generate a graph

Examples

julia> configuration_model([3, 3, 3, 3]; allow_collisions=false)
{4, 6} undirected simple Int64 graph
julia> configuration_model([0, 0])
{2, 0} undirected simple Int64 graph

General.SF_configuration_model — Method

SF_configuration_model(N, γ; min_d=1, allow_collisions=true)

Contruct a graph (of size N) with power-law degree distribution ($P(k)~k^{-γ}$) using the configuration model, with minimal degree min_d.

Examples

julia> SF_configuration_model(4, 2.3; min_d=3, allow_collisions=false)
{4, 6} undirected simple Int64 graph
julia> SF_configuration_model(4, 40)
{4, 2} undirected simple Int64 graph

See also configuration_model

General.spatial_network — Method

spatial_network(ps::Matrix; f=x -> x)

Generate a spatial network with size(ps, 1) nodes embedded in $[0,1]$ ^size(ps,2). Edge $(i,j)$ is present with probability 1-f(norm(ps[i,:] - ps[j,:])/norm(ones(size(ps,2))))

See also Graphs.SimpleGraphs.euclidean_graph

General.spatial_network — Function

g, ps = spatial_network(N[, d=2; f=x -> x])

Generate a spatial network g with N nodes embedded in $[0,1]^d$. ps contain the generated points on which the edges are based.

See also Graphs.SimpleGraphs.euclidean_graph

General.permuted_circle — Method

g = permuted_circle(n, nb_permutations)

Generate a graph with n nodes on a circle (1D periodic grid), i.e. each node is connected to two other nodes. However, the nodes are permuted randomly nb_permutations times, i.e the ordering of the vertices does not correspond to the ordering on the circle anymore.

Examples

julia> g = permuted_circle(4, 1)
{4, 4} undirected simple Int64 graph

General.make_connected! — Method

make_connected!(g::AbstractGraph)

Ensure g only consists of one component by connecting the largest component to each of the other components with one random edge.

See also Graphs.connected_components

Degree distributions

General.degree_distribution — Method

degree_distribution(g::Graph; degree_list=nothing)

Calculate the degree distribution of a given graph g as a tuple (degree_list, degree_counts), such that degree_counts[i] is the number of nodes with degree degree_list[i]. degree_list can be given as an argument, in that case only the degrees in degree_list are considered.

Examples

julia> get_degree_distribution(grid((3, 3, 3), periodic=true))
([6,], [27,])

General.conditional_degree_distribution — Method

conditional_degree_distribution(g::AbstractGraph)

Calculate the conditional degree distribution of a given graph g. The conditional degree distribution is a matrix P where each element P[i,j] represents the probability that a node with degree i is connected to a node with degree j.

General.combined_degree_distribution — Method

combined_degree_distribution(dds::Vector{Tuple{Vector{Int}, Vector{Int}}})

Combine the degree distributions of multiple graphs into one. The input is a vector of tuples, where each tuple contains a degree list and a count list.

General.combined_degree_distribution — Method

combined_degree_distribution(gs::AbstractGraph...)

Combine the degree distributions of multiple graphs into one.

General.joint_distribution — Method

joint_distribution(graphs::AbstractGraph...; count_condition = i -> true)

Calculate the joint degree distribution P ∈ N^{nv(gs[1])^{length(gs)}} for a two layer multiplex network, with each g ∈ gs being a layer. The joint degree distribution is a matrix P where each element P[i1, i2, ...] represents the number of nodes with degree i1 in the first layer, i2 in the second layer, and so on. Only vertices for which count_condition(i) is true are considered.

Graph IO

General.read_from_mtx — Method

read_from_mtx(filename)

Construct a simple graph from a adjacency matrix stored in matrix market format.

See also write_to_mtx

General.write_to_mtx — Method

write_to_mtx(filename)

Store graph g in a matrix market format under filename.

See also read_from_mtx

General.read_from_tsv — Method

read_from_tsv(filename; N::Int64=typemax(Int64), directed=false)

Construct a simple graph from a tsv file. The file should contain two columns with the source and destination of each edge.

Arguments

filename: The path to the file

N: The number of lines to read. If N is smaller than the number of lines in the file, only the first N lines are read.

directed: If true, the graph is directed, otherwise it is undirected.

Examples

julia> g = read_from_tsv("test.tsv")
{4, 3} undirected simple Int64 graph

julia> g = read_from_tsv("test.tsv"; N=2, directed=true)
{4, 2} directed simple Int64 graph

with test.tsv:

1	4
4   1
2	4
3	4

Simplicial complexes

General.vietoris_rips — Method

vietoris_rips(points, epsilon)

Construct the Vietoris-Rips complex of a set of points in Euclidean space.

Arguments

points: A matrix of size d x n where d is the dimension of the Euclidean space and n is the number of points.

epsilon: The radius of the balls used to construct the complex. If epsilon is negative, it will be interpreted as Inf.

General.vietoris_rips_mink — Method

vietoris_rips_mink(n; d=2, min_k=2)

Construct the Vietoris-Rips complex of n random points in d-dimensional Euclidean space. The radius of the balls is chosen such each point is at least part of a min_k dimensional complex.

Current implementation and possibly idea do not make much sense!

General.random_simplicial_complex — Method

random_simplicial_complex(n, p::Vector)

Construct a random simplicial complex on n vertices. The probability of including a simplex of dimension k is p[k], if all of its facets are in the complex. The function will give an AssertionError, if the graph is not connected.