General.jl documentation

General.count_states — Method

count_states(es, i, s)

Compute the number of edges in es which have state s at index i.

Examples

julia> count_states([[1,3], [3,3],[1,1]], 2, 3)
2

General.IStaticEdge — Method

Constructor for an identity StaticEdgewith dimension dim.

See also Ie!, SStaticEdge

General.SStaticEdge — Method

SStaticEdge(selection; invalid=0)

Constructor for a StaticEdge with selection dynamics. The edge has dimension equal to length(selection). After evaluation, e[i]==v_s[i] if selection[i]==true and else e[i]==invalid.

See also combine_graphs, IStaticEdge

General.combine_graphs — Method

static_edges, combined_graph = combine_graphs(gs::AbstractGraph...[; dims=[1], invalid=0])

Generate the combined_graph which edges are given by the union of the edges of gs. The goal of this function is offer an interface that allows to use multiplex networks in network_dynamics. Each argument represents a single layer. The array static_edges consists of SStaticEdges. The named keyword dims gives the dimension the StaticEdge in each layer ought to have. If length(dims)==1, all dimensions are equal to dim[1]. The selection of static_edges[i] is given by

[[edge ∈ edges(gs[i]) for j in 1:dims[i]] for i in eachindex(gs)]

The output of the StaticEdge dynamics corresponding to non-existing edges is set to invalid.

See also SStaticEdge

General.combine_graphs — Method

ode_vertices, static_edges, combined_graph =
    combine_graphs(v_f::Function, gs::AbstractGraph...[; dims=[1]])

Generate the combined_graph which edges are given by the union of the edges of gs. The goal of this function is offer an interface that allows to use multiplex networks in network_dynamics. Each argument represents a single layer. This method ought to bes used when the vertex dynamics depend on the degree in each layer.

Arguments

v_f::Function: a constructor ds -> ODEVertex, where ds[i] is the degree of the corresponding

vertex in gs[i].

gs::AbstractGraph...: graphs to combine

See also SStaticEdge

General.majority_vertex — Method

majority_vertex(i::Integer, dim::Integer; U=1, A=2)

Construct a ODEVertex with binary Watts/threshold dynamics.

Arguments

i::Integer: index of the relevant state

dim::Integer: dimension of the total state

S1=1: value of state 1

S2=2: value of state two

General.SF_configuration_model — Method

SF_configuration_model(N, γ; min_d=1, allow_collisions=true)

Contruct a graph (of size N) with power-law degree distribution ($P(k)~k^{-γ}$) using the configuration model, with minimal degree min_d.

Examples

julia> SF_configuration_model(4, 2.3; min_d=3, allow_collisions=false)
{4, 6} undirected simple Int64 graph
julia> SF_configuration_model(4, 40)
{4, 2} undirected simple Int64 graph

See also configuration_model

General.spatial_network — Method

spatial_network(ps::Matrix; f=x -> x)

Generate a spatial network with size(ps, 1) nodes embedded in $[0,1]$ ^size(ps,2). Edge $(i,j)$ is present with probability 1-f(norm(ps[i,:] - ps[j,:])/norm(ones(size(ps,2))))

See also Graphs.SimpleGraphs.euclidean_graph

General.spatial_network — Function

g, ps = spatial_network(N[, d=2; f=x -> x])

Generate a spatial network g with N nodes embedded in $[0,1]^d$. ps contain the generated points on which the edges are based.

See also Graphs.SimpleGraphs.euclidean_graph

General.read_from_mtx — Method

read_from_mtx(filename)

Construct a simple graph from a adjacency matrix stored in matrix market format.

See also write_to_mtx

General.write_to_mtx — Method

write_to_mtx(filename)

Store graph g in a matrix market format under filename.

See also read_from_mtx

General.make_connected! — Method

make_connected!(g::AbstractGraph)

Ensure g only consists of one component by connecting the largest component to each of the other components with one random edge.

See also Graphs.connected_components